requirements.txt
biopython>=1.81
pandas>=2.0
numpy>=1.24
requests>=2.31
scikit-learn>=1.3
matplotlib>=3.7
plotly>=5.15